All conversions work (still no fractional numbers)
I created a class to organize the functions under one object. Now, they are methods. 

Added, all the functions now output well and this can be paired with a GUI. However, it lacks functionality for fractional digits, which I plan to add later. 

Right now, though, I want to work on a game or a program that takes in API to output something the general public will think is cool, and also help me master Python.